+++ BlackYellow.MVC/Domain/Interfaces/Repositories/ICustomerRepository.cs
@@ -5,6 +5,6 @@ namespace BlackYellow.MVC.Domain.Interfaces.Repositories
     public interface ICustomerRepository : IRepositoryBase<Customer>
     {
         Customer GetCustomerByDocument(string cpf);
-        Customer GetCustomerByUserId(int id);
+        Customer GetCustomerByUserId(long id);
     }
 }
