+++ BlackYellow.MVC/Domain/Interfaces/Services/ICustomerService.cs
@@ -5,6 +5,6 @@ namespace BlackYellow.MVC.Domain.Interfaces.Services
     public interface ICustomerService : IServiceBase<Customer>
     {
         Customer GetCustomerByDocument(string cpf);
-        Customer GetCustomerByUserId(int id);
+        Customer GetCustomerByUserId(long id);
     }
 }
